requirements.txt
absl-py==1.4.0
anyio==4.2.0
apache-beam==2.46.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
avro-python3==1.10.2
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.0
contextlib2==21.6.0
contourpy==1.1.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.8
docopt==0.6.2
etils==1.3.0
exceptiongroup==1.2.0
executing==2.0.1
fastavro==1.9.2
fasteners==0.19
fastjsonschema==2.19.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h5py==3.10.0
hdfs==2.7.3
httplib2==0.21.0
idna==3.6
immutabledict==4.1.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.5.16
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
lvis==0.5.3
lxml==4.9.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mysqlclient==2.2.1
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///C:/laragon/www/kideny-stones/tf_repo/models/research
objsize==0.6.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
portalocker==2.8.2
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.7
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydot==1.4.2
Pygments==2.17.2
pymongo==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.15.0
tensorflow-intel==2.13.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.7.5
tensorflow-text==2.10.0
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
tf-models-official==2.10.1
tf-slim==1.1.0
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typeguard==2.13.3
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
zipp==3.17.0
zstandard==0.22.0
